pyproject.toml
[tool.poetry]
name = "calc_seduc"
version = "0.1.0"
description = ""
authors = ["André Carvalho <alac1984@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = ">=5.2"
pynvim = "^0.4.3"
black = "^22.6.0"
mypy = "^0.961"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
